% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveCalcR.R
\name{compute_rdi}
\alias{compute_rdi}
\title{Compute Regularity Disruption Index (RDI)}
\usage{
compute_rdi(
  activity,
  lambda = 0.005,
  m = 2,
  r = 0.2,
  flow = 1/2000,
  fhigh = 1/300,
  fs = 1
)
}
\arguments{
\item{activity}{Numeric vector of activity values.}

\item{lambda}{Numeric threshold for activity; values below are set to 0. Default is 0.005.}

\item{m}{Integer; the embedding dimension for sample entropy. Default is 2.}

\item{r}{Numeric; the tolerance (similarity) parameter for sample entropy. Default is 0.2.}

\item{flow}{Numeric; low cutoff frequency for the band-pass filter. Default is 1/2000.}

\item{fhigh}{Numeric; high cutoff frequency for the band-pass filter. Default is 1/300.}

\item{fs}{Numeric; sampling frequency. Default is 1.}
}
\value{
A single numeric value representing the RDI.
}
\description{
Calculates the Regularity Disruption Index (RDI) from time series activity data 
using a band-pass filter and sample entropy.
}
