% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveCalcR.R
\name{group_mdv}
\alias{group_mdv}
\title{Add animal group affiliation from the metadata to the main dataset.}
\usage{
group_mdv(data, metadata, animal_col, group_col_meta)
}
\arguments{
\item{data}{MVD table `data.frame`.}

\item{metadata}{MVD metadata table `data.frame`.}

\item{animal_col}{Column name for animal ID in both `data` and `metadata`.}

\item{group_col_meta}{Column name for group information in `metadata`.}
}
\value{
Updated `data.frame` with a new `group` column.
}
\description{
Matches animal IDs between data and metadata and adds a group column.
}
