% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveCalcR.R
\name{mvd_map}
\alias{mvd_map}
\title{Generate a heatmap illustrating movement statistics over time}
\usage{
mvd_map(data, animal_col, time_col, stat_col, group_col, group_sort = NaN)
}
\arguments{
\item{data}{MVD data `data.frame`.}

\item{animal_col}{Column name for animal ID.}

\item{time_col}{Column name for time points.}

\item{stat_col}{Column name for the statistic/value to plot.}

\item{group_col}{Column name for group information.}

\item{group_sort}{Optional character vector to define group order.}
}
\value{
A `ggplot` patchwork object showing the heatmap.
}
\description{
Creates a heatmap of animal data over time with an associated group annotation.
}
