% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveCalcR.R
\name{mvd_rdi}
\alias{mvd_rdi}
\title{Calculate RDI of each animal in MVD data}
\usage{
mvd_rdi(
  data,
  animal_col = "Animal No.",
  group_col = "group",
  groups = NaN,
  stat_col,
  lambda = 0.005,
  m = 2,
  r = 0.2,
  flow = 1/2000,
  fhigh = 1/300,
  fs = 1
)
}
\arguments{
\item{data}{A data frame (MVD) containing the activity data.}

\item{animal_col}{Name of the column indicating animal ID. Default is "Animal No.".}

\item{group_col}{Name of the column indicating experimental group. Default is "group".}

\item{groups}{Optional vector of groups to include; if not specified (default is `NaN`), all groups are used.}

\item{stat_col}{Name of the column with activity values for RDI calculation.}

\item{lambda}{Threshold below which activity is set to zero. Default is 0.005.}

\item{m}{Embedding dimension for sample entropy. Default is 2.}

\item{r}{Tolerance parameter for sample entropy. Default is 0.2.}

\item{flow}{Low cutoff frequency for the filter. Default is 1/2000.}

\item{fhigh}{High cutoff frequency for the filter. Default is 1/300.}

\item{fs}{Sampling frequency. Default is 1.}
}
\value{
A data frame with columns: group, animal, RDI.
}
\description{
Computes the RDI for individual animals across different experimental groups.
}
