% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveCalcR.R
\name{read_mvd_meta}
\alias{read_mvd_meta}
\title{Read MVD (Movement Data) Metadata}
\usage{
read_mvd_meta(path)
}
\arguments{
\item{path}{Path to the MVD file.}
}
\value{
A `data.frame` containing the metadata section.
}
\description{
Extracts metadata from the top section of an MVD file from PhenoMaster device.
}
