% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveCalcR.R
\name{rescale}
\alias{rescale}
\title{Rescale column to [0, 1]}
\usage{
rescale(data, col)
}
\arguments{
\item{data}{A `data.frame`.}

\item{col}{The name of the column to rescale.}
}
\value{
A `data.frame` with the rescaled column.
}
\description{
Linearly rescales a numeric column to the range [0, 1].
}
