% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveCalcR.R
\name{rescale_by_group}
\alias{rescale_by_group}
\title{Rescale values by group}
\usage{
rescale_by_group(data, value_col, group_col)
}
\arguments{
\item{data}{A `data.frame`.}

\item{value_col}{Column name of the values to rescale.}

\item{group_col}{Column name of the grouping variable.}
}
\value{
A `data.frame` with rescaled values per group.
}
\description{
Rescales values in a column within each group to the range [0, 1].
}
