% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveCalcR.R
\name{sample_entropy}
\alias{sample_entropy}
\title{Sample entropy calculation}
\usage{
sample_entropy(series, m = 2, r = 0.2)
}
\arguments{
\item{series}{Numeric vector (time series).}

\item{m}{Embedding dimension.}

\item{r}{Tolerance as a proportion of the standard deviation.}
}
\value{
Numeric value representing sample entropy. Returns `Inf` if undefined.
}
\description{
Calculates the sample entropy of a numeric time series.
}
